requirements.txt
asgiref==3.11.1
Django==6.0.2
djangorestframework==3.16.1
pillow==12.1.1
sqlparse==0.5.5
gunicorn==23.0.0
whitenoise==6.8.2
